More signed compare warnings.

+++ Core/HLE/sceKernelMemory.cpp
@@ -886,7 +886,7 @@ SceUID sceKernelCreateVpl(const char *name, int partition, u32 attr, u32 vplSize
 	// We ignore the upalign to 256 and do it ourselves by 8.
 	u32 allocSize = vplSize;
 	u32 memBlockPtr = userMemory.Alloc(allocSize, (attr & PSP_VPL_ATTR_HIGHMEM) != 0, "VPL");
-	if (memBlockPtr == -1)
+	if (memBlockPtr == (u32)-1)
 	{
 		ERROR_LOG(HLE, "sceKernelCreateVpl: Failed to allocate %i bytes of pool data", vplSize);
 		return SCE_KERNEL_ERROR_NO_MEMORY;
